prettyprinters: simplify Renderable interface
Accurately, split it out into two parts. In the base we don't need to
know if a string is visible, so `Renderable` is now just a wrapper over
`fmt.Stringer`. I've changed the implementation of `String` on
`stringRenderable` to return the proper value if the string is
hidden (an empty string.) Because of this, `Render` is gone as well.

VisibleRenderable takes on the simplifications mentioned in #46. Namely,
it removes `Hide` and `Unhide`. These weren't actually used outside the
internal renderable implementation, so they didn't need to be part of
the interface.

There is some cleanup that had to take place to make this new split
valid, but not too much. Mostly removing utility methods in favor of
just calling `String` and/or wrapping renderables. In basically all
cases, we can just unwrap the string at the call site instead of using a
utility. This may lose a little fidelity, in which case the caller
should definitely be requiring a `VisibleRenderable` instead of just
`Renderable`.

`SprintfVisible` was only needed in one case, it may yet be removed.

Fixes #146