Migrate context for go1.7, refactor fetch/http.go
Go 1.7 was released 2016-08-15, and one of the changes that was
introduced was moving the context library from
`golang.org/x/net/context` to just `context`.

This commit subs in the new package location, but there was one file
that needed extra work: `fetch/http.go`. This file used convenience
functions in `x/net/context/ctxhttp`, which was not included in the new
stdlib version of `context`. I replaced the call to `ctxhttp.Get` with
more common calls, plus a call to `Request.WithContext`.