Refactor from feedback, scale down syntax
There were quite a few critiques in the last commit, so I made edits to
address them. Quite a lot of the concerns were met with renaming
variables and using closures, so the result is that the code is much
more DRY and readable.

The second set of changes came from the fact that HCL does not support
nested list types, even though their README says it does. That threw a
huge monkey wrench into my spec, because I was assuming that nested
lists were possible. And, because time is running out before the feature
freeze, I decided to drop support for the SHOULD level for rules, and
have them be just MUST for now. I will amend the specification
accordingly, before the demo.