Remove unneeded logic, use pkg/errors, redo tests
With @rebeccaskinner's help, we removed logic that was unnecessary for
value passing. With this, it fixed a bad interaction between parameter
validation and value passing that broke rendering. I included new
stanzas to the paramValidation.hcl file that was triggering this bug.

Additionally, I rewrote my param validation code to use the pkg/errors
package instead of the standard errors package and fmt.Errorf. This
makes the errors messages better to read, because I've added context to
all of the err returns.

While I was doing that second thing, I realized that my table test was
really checking for two things: whether or not a given input would
return an error, and also checking the content of the error messages. I
wanted to make sure that the error messages fit with the specification,
but that can be checked with a different test. So, I wrote a stub test
that will check the error messages, and I rewrote the existing table
test to only check if something returns an error or not.